검색기능 오류 해결, gacha기능 추가 및 랜덤 포켓몬 조회 기능 추가 DB연결 및 iframe 구현중

+++ src/main/java/org/choongang/pokemon/entities/PokemonDetail.java
@@ -11,7 +11,7 @@ public class PokemonDetail extends AbstractPokemonDetail {
     private String name;
     private String nameKr;
     private int weight;
-    private int height;
+    private double height;
     private int baseExperience;
     private String frontImage;
     private String backImage;
